# api-contracts/openapi/ctmp.openapi.yaml
openapi: 3.0.3
info:
  title: CTMP API
  version: 1.0.0
  description: >
    Corporate Tender Management Platform API contract. Commercial envelope state
    and commercial visibility are separate controls: committee opening changes
    envelope state only, while commercial details still require explicit
    commercial permissions. All sensitive state changes, views, downloads,
    exports, permission changes, and exceptions must be audit logged.
servers:
  - url: /api/v1
    description: Versioned API root
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Internal user login
      description: Authenticates internal users, typically backed by Active Directory.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh an access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
  /auth/mfa/verify:
    post:
      tags: [Auth]
      summary: Verify internal MFA challenge
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaVerifyRequest'
      responses:
        '200':
          description: MFA verified.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
  /vendor-auth/register:
    post:
      tags: [Vendor Auth]
      summary: Register a vendor organization
      description: >
        Public registration must validate CAPTCHA server-side, rate limit the
        request, log suspicious attempts, create an email verification token, and
        avoid creating an approved vendor before internal review.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorRegistrationRequest'
      responses:
        '202':
          description: Registration accepted for verification and review.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorRegistrationAccepted'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
  /vendor-auth/login:
    post:
      tags: [Vendor Auth]
      summary: Vendor user login
      description: >
        Authenticates vendor users by email and password. Returns an access
        token or, when MFA is enabled for the account, a challengeId for use
        with POST /vendor-auth/mfa/verify.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorLoginRequest'
      responses:
        '200':
          description: Login succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /vendor-auth/verify-email:
    post:
      tags: [Vendor Auth]
      summary: Verify vendor email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '204':
          description: Email verified.
  /vendor-auth/forgot-password:
    post:
      tags: [Vendor Auth]
      summary: Request vendor password reset
      description: Uses CAPTCHA and rate limiting and must not reveal whether the email exists.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '202':
          description: Request accepted.
  /vendor-auth/reset-password:
    post:
      tags: [Vendor Auth]
      summary: Reset vendor password
      description: Reset token is single-use, expiring, audited, and invalidates active vendor sessions.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '204':
          description: Password reset completed.
  /vendor-auth/me:
    get:
      tags: [Vendor Auth]
      summary: Get the calling vendor's profile (self-service)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Vendor profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorProfileResponse'
    patch:
      tags: [Vendor Auth]
      summary: Update non-sensitive fields on the calling vendor's profile
      description: Email change is rejected here — use the verified-email flow.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorProfileUpdateRequest'
      responses:
        '200':
          description: Updated vendor profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorProfileResponse'
  /vendor-auth/me/bids:
    get:
      tags: [Vendor Auth]
      summary: List the calling vendor's bids across all tenders
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Vendor's bids.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyBidsListResponse'
  /vendor-auth/mfa/verify:
    post:
      tags: [Vendor Auth]
      summary: Verify vendor MFA challenge
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaVerifyRequest'
      responses:
        '200':
          description: MFA verified.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
...
  /vendors:
    get:
      tags: [Vendors]
      summary: List vendors (admin)
      description: Requires vendors:list. Filter by registrationStatus.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING_APPROVAL, APPROVED, REJECTED, SUSPENDED, BLACKLISTED]
      responses:
        '200':
          description: Vendor list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorListResponse'
  /vendors/{vendorId}:
    get:
      tags: [Vendors]
      summary: Get vendor profile
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/VendorId'
      responses:
        '200':
          description: Vendor profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vendor'
    patch:
      tags: [Vendors]
      summary: Update vendor profile
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/VendorId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorUpdateRequest'
      responses:
        '200':
          description: Vendor updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vendor'
  /vendors/{vendorId}/approve:
    post:
      tags: [Vendors]
      summary: Approve vendor registration
      description: Requires vendors:approve. Vendor must have emailVerified=true. Transitions PENDING_APPROVAL → APPROVED. Sends approval email. Audit logged.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/VendorId'
      responses:
        '200':
          description: Vendor approved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vendor'
  /vendors/{vendorId}/reject:
    post:
      tags: [Vendors]
      summary: Reject vendor registration
      description: Reason required for audit. Transitions PENDING_APPROVAL → REJECTED.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/VendorId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Vendor rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vendor'
  /vendors/{vendorId}/suspend:
    post:
      tags: [Vendors]
      summary: Suspend approved vendor
      description: Reason required for audit. Transitions APPROVED → SUSPENDED. Revokes active vendor sessions (bumps token version).
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/VendorId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Vendor suspended.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vendor'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    VendorId:
      name: vendorId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
  responses:
    Unauthorized:
      description: Authentication failed or token is missing.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: Too many requests.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: CTMP_VALIDATION_ERROR
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            traceId:
              type: string
    VendorLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    MfaVerifyRequest:
      type: object
      required: [challengeId, code]
      properties:
        challengeId:
          type: string
        code:
          type: string
    AuthTokenResponse:
      type: object
      required: [accessToken, tokenType, expiresIn]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
    TokenRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
    ForgotPasswordRequest:
      type: object
      required: [email, captchaToken]
      properties:
        email:
          type: string
          format: email
        captchaToken:
          type: string
    ResetPasswordRequest:
      type: object
      required: [token, newPassword, captchaToken]
      properties:
        token:
          type: string
        newPassword:
          type: string
          format: password
        captchaToken:
          type: string
    VendorRegistrationRequest:
      type: object
      required: [companyName, registrationNumber, contactName, email, password, captchaToken]
      properties:
        companyName:
          type: string
        registrationNumber:
          type: string
        taxNumber:
          type: string
        contactName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          format: password
        captchaToken:
          type: string
    VendorRegistrationAccepted:
      type: object
      required: [registrationRequestId, status]
      properties:
        registrationRequestId:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING_EMAIL_VERIFICATION, PENDING_REVIEW]
    TenderStatus:
      type: string
      enum:
        - DRAFT
        - INTERNAL_REVIEW
        - APPROVED
        - PUBLISHED
        - CLARIFICATION_PERIOD
        - SUBMISSION_CLOSED
        - TECHNICAL_OPENING
        - TECHNICAL_EVALUATION
        - COMMERCIAL_SEALED
        - COMMITTEE_COMMERCIAL_OPENING
        - COMMERCIAL_EVALUATION
        - AWARD_RECOMMENDATION
        - AWARDED
        - TENDER_CLOSED
        - CANCELLED
        - SUSPENDED
        - ARCHIVED
...
    EnvelopeStatus:
      type: string
      enum: [DRAFT, SUBMITTED, SEALED, OPENED, LOCKED]
    BidStatus:
      type: string
      enum: [DRAFT, SUBMITTED, LATE_SUBMITTED, LATE_ACCEPTED, WITHDRAWN, DISQUALIFIED]
...
    Vendor:
      type: object
      required: [id, company, email, contactName, registrationStatus, emailVerified, registeredAt]
      properties:
        id:
          type: string
          format: uuid
        company:
          type: string
        email:
          type: string
        contactName:
          type: string
        contactPhone:
          type: string
        country:
          type: string
        category:
          type: string
        taxId:
          type: string
        registrationStatus:
          type: string
          enum: [PENDING_APPROVAL, APPROVED, REJECTED, SUSPENDED, BLACKLISTED]
        emailVerified:
          type: boolean
        registeredAt:
          type: string
          format: date-time
        approvedAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
        documentCount:
          type: integer
    VendorListResponse:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Vendor'
    VendorUpdateRequest:
      type: object
      properties:
        company:
          type: string
        contactName:
          type: string
        contactPhone:
          type: string
        country:
          type: string
        category:
          type: string
        taxId:
          type: string
...
    VendorProfileResponse:
      type: object
      required: [vendor]
      properties:
        vendor:
          type: object
          required: [id, companyName, status, registeredAt]
          properties:
            id:
              type: string
              format: uuid
            companyName:
              type: string
            registrationNumber:
              type: string
            taxNumber:
              type: string
            country:
              type: string
            address:
              type: string
            phone:
              type: string
            website:
              type: string
            status:
              type: string
              enum: [PENDING, APPROVED, REJECTED, SUSPENDED, BLACKLISTED]
            registeredAt:
              type: string
              format: date-time
            approvedAt:
              type: string
              format: date-time
              nullable: true
        primaryContact:
          type: object
          nullable: true
          required: [id, email, fullName, emailVerified, mfaEnabled]
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
            fullName:
              type: string
            phone:
              type: string
            emailVerified:
              type: boolean
            lastLoginAt:
              type: string
              format: date-time
              nullable: true
            mfaEnabled:
              type: boolean
    VendorProfileUpdateRequest:
      type: object
      properties:
        companyName:
          type: string
        taxNumber:
          type: string
        country:
          type: string
        address:
          type: string
        phone:
          type: string
        website:
          type: string
        contactFullName:
          type: string
        contactPhone:
          type: string
    MyBidsListResponse:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/MyBidSummary'
    MyBidSummary:
      type: object
      required: [id, tenderId, tenderReference, tenderTitle, status, technicalEnvelopeStatus, commercialEnvelopeStatus]
      properties:
        id:
          type: string
          format: uuid
        tenderId:
          type: string
          format: uuid
        tenderReference:
          type: string
        tenderTitle:
          type: string
        tenderStatus:
          $ref: '#/components/schemas/TenderStatus'
        submissionDeadline:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: '#/components/schemas/BidStatus'
        submittedAt:
          type: string
          format: date-time
          nullable: true
        technicalResult:
          type: string
          enum: [PASS, FAIL]
        technicalEnvelopeStatus:
          $ref: '#/components/schemas/EnvelopeStatus'
        commercialEnvelopeStatus:
          $ref: '#/components/schemas/EnvelopeStatus'
        receiptNumber:
          type: string
...
    PaginatedResponse:
      type: object
      required: [page, pageSize, total]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer